+++ HISTORY.rst
@@ -10,6 +10,7 @@ Unreleased
 
 - Added support for ``IA_CONFIG_FILE`` environment variable to specify the configuration file path.
 - Added ``--no-derive`` option to ``ia copy`` and ``ia move``.
+- Added ``--no-backup`` option to ``ia copy``, ``ia move``, ``ia upload``, and ``ia delete``.
 
 2.2.0 (2021-11-23)
 ++++++++++++++++++